Speed up encoding with fixed-size flushes and paired literals
BitWriter flushed full bytes through a variable-length extend_from_slice,
which called memcpy on every flush. It now appends all 8 buffer bytes and
truncates to the full-byte count, which compiles to one store.

The q2 to q5 token writers emitted literals one at a time. They now use
the paired literal emitter from q1. Length and distance extra bits are
stored as u32, which shrinks PreparedToken.

Encoding is 0.6% to 38% faster, most at q0 and on low-compressibility
data. Output is byte-identical to the previous commit.

+++ crates/encode/src/encode/q0.rs
@@ -111,9 +111,9 @@ impl Batch {
         self.records.push(Record {
             insert_start: token.insert_start as u32,
             insert_len: token.insert_len as u32,
-            insert_extra: insert.extra as u32,
-            copy_extra: copy.map_or(0, |copy| copy.extra as u32),
-            distance_extra: distance.map_or(0, |distance| distance.extra as u32),
+            insert_extra: insert.extra,
+            copy_extra: copy.map_or(0, |copy| copy.extra),
+            distance_extra: distance.map_or(0, |distance| distance.extra),
             meta: pack_record_meta(
                 u32::from(command_symbol),
                 distance_symbol,
